add auto_login&directory
