+++ lib/Pages/PostPage.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+
+class PostPage extends StatefulWidget {
+  const PostPage({Key? key}) : super(key: key);
+
+  @override
+  _PostPageState createState() => _PostPageState();
+}
+
+class _PostPageState extends State<PostPage> {
+  @override
+  Widget build(BuildContext context) {
+    return const Scaffold(
+      body: Center(
+        child: Text('postPage'),
+      ),
+    );
+  }
+}
