+++ lib/main.dart
@@ -1,3 +1,5 @@
+import 'package:chinese_sns/Pages/PostPage.dart';
+import 'package:chinese_sns/Pages/SingInPage.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/cupertino.dart';
@@ -22,151 +24,11 @@ class MyApp extends StatelessWidget {
         }
         if (snapshot.hasData) {
           // User が null でなない、つまりサインイン済みのホーム画面へ
-          return SignInPage();
+          return const PostPage();
         }
         // User が null である、つまり未サインインのサインイン画面へ
-        return HomePage();
+        return const SignInPage();
       },
     ),
   );
 }
-
-class SignInPage extends StatefulWidget {
-  @override
-  const SignInPage({Key? key}) : super(key: key);
-  _SignInPage createState() => _SignInPage();
-}
-
-class _SignInPage extends State<SignInPage> {
-  String Email = "";
-  String Password = "";
-
-  // 登録・ログインに関する情報を表示
-  String infoText = "";
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Container(
-          padding: EdgeInsets.all(32),
-          child: Column(
-            mainAxisSize: MainAxisSize.min, //カラムの位置を調整できるように軸方向のサイズを最小に
-            children: <Widget>[
-              TextFormField(
-                decoration: InputDecoration(labelText: "メールアドレス"),
-                onChanged: (String value) {
-                  setState(() {
-                    Email = value;
-                  });
-                },
-              ),
-              TextFormField(
-                decoration: InputDecoration(labelText: "パスワード"),
-                obscureText: true,
-                onChanged: (String value) {
-                  setState(() {
-                    Password = value;
-                  });
-                },
-              ),
-              const SizedBox(height: 8),
-              ElevatedButton(
-                onPressed: () {
-                  _showModalPicker(context);
-                },
-                child: Text(_selectedItem),
-              ),
-
-              const SizedBox(height: 8),
-              ElevatedButton(
-                onPressed: () async {
-                  try {
-                    // メール/パスワードでログイン
-                    final FirebaseAuth auth = FirebaseAuth.instance;
-                    final UserCredential result =
-                    await auth.signInWithEmailAndPassword(
-                      email: Email,
-                      password: Password,
-                    );
-                    // ログインに成功した場合
-                    final User user = result.user!;
-                    setState(() {
-                      infoText = "ログインOK：${user.email}";
-                    });
-                  } catch (e) {
-                    // ログインに失敗した場合
-                    setState(() {
-                      infoText = "ログインNG：${e.toString()}";
-                    });
-                  }
-                },
-                child: const Text("Login"),
-                style: ElevatedButton.styleFrom(
-                  primary: Colors.orangeAccent, //ボタンの背景色
-                ),
-              ),
-              const SizedBox(height: 8),
-              Text(infoText),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-
-  void _showModalPicker(BuildContext context) {
-    showModalBottomSheet<void>(
-      context: context,
-      builder: (BuildContext context) {
-        return Container(
-          height: MediaQuery
-              .of(context)
-              .size
-              .height / 3,
-          child: GestureDetector(
-            onTap: () {
-              Navigator.pop(context);
-            },
-            child: CupertinoPicker(
-              itemExtent: 40,
-              children: _items.map(_pickerItem).toList(),
-              onSelectedItemChanged: _onSelectedItemChanged,
-            ),
-          ),
-        );
-      },
-    );
-  }
-
-  String _selectedItem = 'Teacher or Student';
-
-  final List<String> _items = [
-    'Teacher',
-    'Student',
-  ];
-
-  Widget _pickerItem(String str) {
-    return Text(
-      str,
-      style: const TextStyle(fontSize: 32),
-    );
-  }
-
-  void _onSelectedItemChanged(int index) {
-    setState(() {
-      _selectedItem = _items[index];
-    });
-  }
-}
-
-class HomePage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) => const Scaffold(
-    body: Center(
-      child: Text('サインイン済み時に表示するホーム画面'),
-    ),
-  );
-}
-
